WIP on a timer widget.
Currently there's a timer that's Not Exactly Ready For Prime Time.
It needs, at minimum, better timekeeping (instead of "Always ask for
a 1-second timeout, ignoring processing lag") and some kind of alert
when the time is up for it to be useful for anything.
It also needs a prettier interface and more-featureful controls to
be actually worth using.